+++ docs/griptape-framework/drivers/src/event_listener_drivers_4.py
@@ -3,12 +3,12 @@ import os
 from griptape.configs import Defaults
 from griptape.configs.drivers import DriversConfig
 from griptape.drivers import AwsIotCoreEventListenerDriver, OpenAiChatPromptDriver
-from griptape.events import EventListener, FinishStructureRunEvent, event_bus
+from griptape.events import EventBus, EventListener, FinishStructureRunEvent
 from griptape.rules import Rule
 from griptape.structures import Agent
 
 Defaults.drivers_config = DriversConfig(prompt_driver=OpenAiChatPromptDriver(model="gpt-3.5-turbo", temperature=0.7))
-event_bus.add_event_listeners(
+EventBus.add_event_listeners(
     [
         EventListener(
             event_types=[FinishStructureRunEvent],