+++ griptape/configs/defaults_config.py
@@ -4,6 +4,8 @@ from typing import TYPE_CHECKING
 
 from attrs import Factory, define, field
 
+from griptape.mixins.singleton_mixin import SingletonMixin
+
 from .base_config import BaseConfig
 from .drivers.openai_drivers_config import OpenAiDriversConfig
 from .logging.logging_config import LoggingConfig
@@ -13,14 +15,7 @@ if TYPE_CHECKING:
 
 
 @define(kw_only=True)
-class _DefaultsConfig(BaseConfig):
-    _instance = None
-
-    def __new__(cls, *args, **kwargs) -> _DefaultsConfig:
-        if not cls._instance:
-            cls._instance = super().__new__(cls, *args, **kwargs)  # noqa: UP008
-        return cls._instance
-
+class _DefaultsConfig(BaseConfig, SingletonMixin):
     logging_config: LoggingConfig = field(default=Factory(lambda: LoggingConfig()))
     drivers_config: BaseDriversConfig = field(default=Factory(lambda: OpenAiDriversConfig()))