+++ griptape/mixins/__init__.py
@@ -5,6 +5,7 @@ from .rule_mixin import RuleMixin
 from .serializable_mixin import SerializableMixin
 from .media_artifact_file_output_mixin import BlobArtifactFileOutputMixin
 from .futures_executor_mixin import FuturesExecutorMixin
+from .singleton_mixin import SingletonMixin
 
 __all__ = [
     "ActivityMixin",
@@ -14,4 +15,5 @@ __all__ = [
     "BlobArtifactFileOutputMixin",
     "SerializableMixin",
     "FuturesExecutorMixin",
+    "SingletonMixin",
 ]