+++ griptape/tasks/base_task.py
@@ -10,7 +10,7 @@ from attrs import Factory, define, field
 
 from griptape.artifacts import ErrorArtifact
 from griptape.configs import Defaults
-from griptape.events import FinishTaskEvent, StartTaskEvent, event_bus
+from griptape.events import EventBus, FinishTaskEvent, StartTaskEvent
 from griptape.mixins import FuturesExecutorMixin
 
 if TYPE_CHECKING:
@@ -137,7 +137,7 @@ class BaseTask(FuturesExecutorMixin, ABC):
 
     def before_run(self) -> None:
         if self.structure is not None:
-            event_bus.publish_event(
+            EventBus.publish_event(
                 StartTaskEvent(
                     task_id=self.id,
                     task_parent_ids=self.parent_ids,
@@ -149,7 +149,7 @@ class BaseTask(FuturesExecutorMixin, ABC):
 
     def after_run(self) -> None:
         if self.structure is not None:
-            event_bus.publish_event(
+            EventBus.publish_event(
                 FinishTaskEvent(
                     task_id=self.id,
                     task_parent_ids=self.parent_ids,