+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Parameter `structure` to `BaseTask`.
 - Method `try_find_task` to `Structure`.
 - `TranslateQueryRagModule` `RagEngine` module for translating input queries.
-- Global event bus, `griptape.events.event_bus`, for publishing and subscribing to events.
+- Global event bus, `griptape.events.EventBus`, for publishing and subscribing to events.
 - Global object, `griptape.configs.Defaults`, for setting default values throughout the framework.
 - Unique name generation for all `RagEngine` modules.
 - `ExtractionTool` for having the LLM extract structured data from text.
@@ -27,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `griptape.utils.decorators.lazy_property` for creating lazy properties.
 
 ### Changed
-- **BREAKING**: Removed all uses of `EventPublisherMixin` in favor of `event_bus`.
+- **BREAKING**: Removed all uses of `EventPublisherMixin` in favor of `EventBus`.
 - **BREAKING**: Removed `EventPublisherMixin`.
 - **BREAKING**: Removed `Pipeline.prompt_driver` and `Workflow.prompt_driver`. Set this via `griptape.configs.Defaults.drivers.prompt_driver` instead. `Agent.prompt_driver` has not been removed.
 - **BREAKING**: Removed `Pipeline.stream` and `Workflow.stream`. Set this via `griptape.configs.Defaults.drivers.prompt_driver.stream` instead. `Agent.stream` has not been removed.